🎨 index.js is now splitted
