+++ src/handlers/index.js
@@ -1,0 +1,5 @@
+const events = require('./events');
+const commands = require('./commands');
+const locale = require('./locale');
+
+module.exports = { events, commands, locale };
