+++ src/helpers/index.js
@@ -1,0 +1,4 @@
+const database = require('./database');
+const deployCommands = require('./deployCommands');
+
+module.exports = { database, deployCommands };
